requirements.txt
behave==1.2.6
requests==2.32.4
PyHamCrest==2.0.2
allure-behave==2.15.0
allure-python-commons==2.9.43
html-testRunner==1.2.1
webdriver-manager==4.0.2
selenium


